+++ shared/database/models/tariff.py
@@ -12,8 +12,8 @@ class Tariff(BaseModel):
     name: Mapped[str] = mapped_column(unique=True)
     description: Mapped[str] = mapped_column()
     duration: Mapped[timedelta] = mapped_column()
-    price: Mapped[float] = mapped_column()
-    price_of_traffic_reset: Mapped[float] = mapped_column()
+    price: Mapped[int] = mapped_column()
+    price_of_traffic_reset: Mapped[int] = mapped_column()
     traffic: Mapped[int] = mapped_column()
 
     with_access: Mapped[bool] = mapped_column()
